How to run the file
What you're supposed to do. is to either create a text file, or manually add words to the "passwords" array. Then is is reccomended that you run the code out of a terminal, but you may still run it straight up. This code will work on Windows, Mac, and Linux.